+++ src/components/content/care-team/patient-careteam.tsx
@@ -7,7 +7,6 @@ import { Table } from "@/components/core/table/table";
 import { usePatientCareTeam } from "@/fhir/care-team";
 import { CareTeamPractitionerModel } from "@/fhir/models/careteam-practitioner";
 import { useBreakpoints } from "@/hooks/use-breakpoints";
-import { useFQSFeatureToggle } from "@/hooks/use-feature-toggle";
 
 export type PatientCareTeamProps = {
   className?: string;
@@ -21,7 +20,6 @@ export type CareTeamDetailsDrawerProps = {
 };
 
 export function PatientCareTeam({ className }: PatientCareTeamProps) {
-  const { enabled } = useFQSFeatureToggle("careTeams");
   const containerRef = useRef<HTMLDivElement>(null);
   const breakpoints = useBreakpoints(containerRef);
   const patientCareTeamQuery = usePatientCareTeam();
@@ -31,7 +29,6 @@ export function PatientCareTeam({ className }: PatientCareTeamProps) {
     subHeader: (m) => m.qualification,
     details: careTeamData,
     getSourceDocument: true,
-    enableFQS: enabled,
   });
 
   return (